staging: all changes on top of upstream/master

+++ PKGBUILD
@@ -27,29 +27,25 @@ plain '             `.-:///////:-.`'
 _where="$PWD" # track basedir as different Arch based distros are moving srcdir around
 
 source "$_where"/linux-tkg-config/prepare
+source "$_where"/linux-tkg-config/staging # This will source the staging files for the 'knoobs' configs
 
 # Create BIG_UGLY_FROGMINER only on first run and save in it all settings
-if [ ! -e "$_where"/BIG_UGLY_FROGMINER ]; then
+_frogminer_bootstrap "${_where}/BIG_UGLY_FROGMINER" "${_where}/BIG_UGLY_FROGMINER.pending"
 
-  aggregate_user_config
-
-  echo -e "_ispkgbuild=\"true\"\n_distro=\"Arch\"\n_where=\"$PWD\"" >> "$_where"/BIG_UGLY_FROGMINER
-
-  source "$_where"/BIG_UGLY_FROGMINER
-
-  _tkg_initscript
-fi
-
-source "$_where"/BIG_UGLY_FROGMINER
-
+_vanilla_tag=""
+_vanilla_mode && _vanilla_tag="vanilla-"
 if [ -n "$_custom_pkgbase" ]; then
   pkgbase="${_custom_pkgbase}"
 else
-  pkgbase=linux"${_basever}"-tkg-"${_cpusched}"${_compiler_name}
+  pkgbase="linux${_basever}-tkg-${_vanilla_tag}${_cpusched}${_compiler_name}"
 fi
 pkgname=("${pkgbase}" "${pkgbase}-headers")
+while IFS= read -r _pkgname_extra; do
+  [ -n "$_pkgname_extra" ] && pkgname+=("$_pkgname_extra")
+done < <(_run_staging_knoob _pkgname_extra)
+unset _pkgname_extra
 pkgver="${_basekernel}"."${_sub}"
-pkgrel=273
+pkgrel=8086
 pkgdesc='Linux-tkg'
 arch=('x86_64') # no i686 in here
 url="https://www.kernel.org/"
@@ -78,8 +74,19 @@ makedepends=(
 if [ "$_compiler_name" = "-llvm" ]; then
   makedepends+=('clang' 'llvm' 'lld')
 fi
+while IFS= read -r _makedep_extra; do
+  [ -n "$_makedep_extra" ] && makedepends+=("$_makedep_extra")
+done < <(_run_staging_knoob _makedepends_extra)
+unset _makedep_extra
 optdepends=('schedtool')
 options=('!strip' 'docs')
+
+while IFS= read -r _source_extra; do
+  [ -n "$_source_extra" ] || continue
+  source+=("$_source_extra")
+  sha256sums+=("SKIP")
+done < <(_run_staging_knoob _source_extra)
+unset _source_extra
 
 for f in "$_where"/linux-tkg-config/"$_basekernel"/* "$_where"/linux-tkg-patches/"$_basekernel"/*.patch; do
   source+=( "$f" )
@@ -97,6 +104,8 @@ prepare() {
   rm -rf $pkgdir # Nuke the entire pkg folder so it'll get regenerated clean on next build
 
   ln -sf "${_kernel_work_folder_abs}" "${srcdir}"
+
+  _run_staging_knoob _prepare_after_link
 
   _tkg_srcprep
 }
@@ -151,9 +160,12 @@ build() {
     export KCFLAGS
     export KRUSTFLAGS
 
-    time ( make ${_force_all_threads} ${llvm_opt} LOCALVERSION= bzImage modules 2>&1 ) 3>&1 1>&2 2>&3
+    _make_extra_flags="$(_run_staging_knoob _build_make_flags)"
+    time ( make ${_force_all_threads} ${llvm_opt} ${_make_extra_flags} LOCALVERSION= bzImage modules 2>&1 ) 3>&1 1>&2 2>&3
     return 0
   )
+
+  _run_staging_knoob _build_after_kernel
 }
 
 hackbase() {
@@ -161,10 +173,13 @@ hackbase() {
 
   pkgdesc="The $pkgdesc kernel and modules - https://github.com/Frogging-Family/linux-tkg"
   depends=('coreutils' 'kmod' 'initramfs')
+  local _docs_optdepend
+  _docs_optdepend="$(_run_staging_knoob _docs_optdepend)"
+  [ -n "$_docs_optdepend" ] || _docs_optdepend='linux-docs: Kernel hackers manual - HTML documentation that comes with the Linux kernel.'
   optdepends=(
     "$pkgname-headers: headers and scripts for building modules"
     'linux-firmware: firmware images needed for some devices'
-    'linux-docs: Kernel hackers manual - HTML documentation that comes with the Linux kernel.'
+    "$_docs_optdepend"
     'nvidia-tkg: NVIDIA drivers for all installed kernels - non-dkms version.'
     'nvidia-dkms-tkg: NVIDIA drivers for all installed kernels - dkms version.'
     'modprobed-db: Keeps track of EVERY kernel module that has ever been probed. Useful for make localmodconfig.'
@@ -198,8 +213,10 @@ hackbase() {
   local _STRIP_MODS=""
   [[ "$_STRIP" == "true" ]] && _STRIP_MODS="INSTALL_MOD_STRIP=1"
 
-  ZSTD_CLEVEL=19 make INSTALL_MOD_PATH="$pkgdir/usr" $_STRIP_MODS \
+  ZSTD_CLEVEL=19 make ${llvm_opt} INSTALL_MOD_PATH="$pkgdir/usr" $_STRIP_MODS \
     DEPMOD=/doesnt/exist modules_install  # Suppress depmod
+
+  _run_staging_knoob _pkgbase_after_modules_install "${modulesdir}"
 
   # remove build and source links
   rm -f "$modulesdir"/{source,build}
@@ -237,7 +254,7 @@ hackheaders() {
   source "$_where"/BIG_UGLY_FROGMINER
 
   pkgdesc="Headers and scripts for building modules for the $pkgdesc kernel - https://github.com/Frogging-Family/linux-tkg"
-  provides=("linux-headers=${pkgver}" "${pkgbase}-headers=${pkgver}")
+  provides=("linux-headers=${pkgver}" "${pkgbase}-headers=${pkgver}" LINUX-HEADERS)
   depends=(
     binutils
     glibc
@@ -299,8 +316,12 @@ hackheaders() {
     rm -r "$arch"
   done
 
-  msg2 "Removing documentation..."
-  rm -r "$builddir/Documentation"
+  if declare -F _headers_keep_docs >/dev/null && _headers_keep_docs; then
+    :
+  else
+    msg2 "Removing documentation..."
+    rm -r "$builddir/Documentation"
+  fi
 
   msg2 "Removing broken symlinks..."
   find -L "$builddir" -type l -printf 'Removing %P\n' -delete
@@ -327,12 +348,14 @@ hackheaders() {
   mkdir -p "$pkgdir/usr/src"
   ln -sr "$builddir" "$pkgdir/usr/src/$pkgbase"
 
+  _run_staging_knoob _headers_after_install "${builddir}"
+
   if [ "$_STRIP" = "true" ]; then
     echo "Stripping vmlinux..."
     strip -v $STRIP_STATIC "$builddir/vmlinux"
   fi
 
-  if [ "$_NUKR" = "true" ]; then
+  if [ "$_NUKR" = "true" ] && _run_staging_knoob _headers_should_cleanup_srcdir; then
     rm -rf "$srcdir" # Nuke the entire src folder so it'll get regenerated clean on next build
   fi
 }
@@ -345,4 +368,6 @@ hackbase
 package_${pkgbase}-headers() {
 hackheaders
 }
+
+$(_run_staging_knoob _package_functions)
 EOF
